#84 - updated interface for fetching tournament details

+++ src/main/kotlin/pl/cyganki/utils/model/TaskUserDetails.kt
@@ -2,7 +2,7 @@ package pl.cyganki.utils.model
 
 data class TaskUserDetails(
         val taskId: String,
-        val earnedPoints: Long,
+        val earnedPoints: Int,
         val userAttempts: Int,
         val maxAttempts: Int?
 )
